+++ src/components/button/index.js
@@ -1,6 +1,5 @@
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
-import SplitButton from './SplitButton';
 
 export default Button;
-export { Button, ButtonGroup, SplitButton };
+export { Button, ButtonGroup };